Fixed a text error by adding an else statement; might be done soon
When inputting a word that isn't a keyword, it will originally lead a bug that printed incomplete responses from the text file.  This fixes the problem, if you input a word that isn't a keyword in the project.  Also, I might also be close to completion, so this will be submitted soon.